+++ src/main/java/jm/task/core/jdbc/model/User.java
@@ -22,10 +22,10 @@ public class User {
 
     }
 
-    public User(String name, String lastName, Byte age) {
+    public User(String name, String lastName, int age) {
         this.name = name;
         this.lastName = lastName;
-        this.age = age;
+        this.age = (byte) age;
     }
 
     public Long getId() {
@@ -56,8 +56,8 @@ public class User {
         return age;
     }
 
-    public void setAge(Byte age) {
-        this.age = age;
+    public void setAge(int age) {
+        this.age = (byte) age;
     }
 
     @Override
